+++ src/main/java/br/com/postech/software/architecture/techchallenge/repository/jpa/ProdutoJpaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProdutoJpaRepository extends JpaRepository<Produto, Integer> {
+public interface ProdutoJpaRepository extends JpaRepository<Produto, Long> {
 
     List<Produto> findByCategoria(CategoriaEnum categoria);
